Only delete files/folders if they exist on filesystem
This broke "add a folder" in manage-files.feature

+++ src/SilverStripe/BehatExtension/Context/BasicContext.php
@@ -205,7 +205,7 @@ JS;
     public function cleanAssetsAfterScenario(ScenarioEvent $event)
     {
         foreach(\File::get() as $file) {
-            $file->delete();
+            if(file_exists($file->getFullPath())) $file->delete();
         }
     }
